# src/AspNetCoreSample.WebApi/OpenAPI/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "AspNetCoreSample WebApi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Token/Auth": {
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "Keycloakによる認証を行う",
        "operationId": "Token_Auth",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "ユーザ名およびパスワード",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "AccessTokenおよびRefreshTokenの取得",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Token/RefreshToken": {
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "RefreshTokenを使用してAccessTokenを更新する",
        "operationId": "Token_RefreshToken",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "RefreshToken",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "AccessTokenおよびRefreshTokenの取得",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Token/RevokeToken": {
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "RefreshTokenを無効化する",
        "operationId": "Token_RevokeToken",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": true,
            "description": "Language preference for the response.",
            "schema": {
              "type": "string",
              "items": {
                "type": "string"
              },
              "enum": [
                "ja-JP",
                "en-US"
              ]
            },
            "default": "ja-JP",
            "nullable": true
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "RefreshToken",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "なし",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userName",
          "password"
        ],
        "properties": {
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UpdateTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "RevokeTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "description": "Bearer {token}",
        "in": "header",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}